refactor: removes returns and updates providers
