+++ src/providers/openai.ts
@@ -1,16 +1,17 @@
 import { ChatGPTAPI } from "chatgpt";
 import { Configuration, OpenAIApi } from "openai";
-import config from "../config";
+
+import { config } from "../env/env-variables";
 
 // ChatGPT Client (text-davinci-003)
 export const chatgpt = new ChatGPTAPI({
-	apiKey: config.openAIAPIKey,
-	maxModelTokens: config.maxModelTokens
+	apiKey: config.OPENAI_API_KEY,
+	maxModelTokens: config.MAX_MODEL_TOKENS
 });
 
 // OpenAI Client (DALL-E)
 export const openai = new OpenAIApi(
 	new Configuration({
-		apiKey: config.openAIAPIKey
+		apiKey: config.OPENAI_API_KEY
 	})
 );
